requirements.txt
# PM Updown Bot dependencies — pinned from pip list
# Run: pip install -r requirements.txt
aiohttp==3.13.3
aiosignal==1.4.0
anyio==4.12.1
apify_client==2.5.0
bcrypt==3.2.2
beautifulsoup4==4.14.3
bitarray==3.8.0
ccxt==4.5.40
certifi==2026.1.4
cffi==2.0.0
click==8.1.6
coincurve==21.0.0
cryptography==41.0.7
cytoolz==1.1.0
eth_abi==5.2.0
Flask==3.1.3
frozenlist==1.8.0
h11==0.16.0
hexbytes==1.3.1
httpx==0.28.1
idna==3.6
multidict==6.7.1
numpy==2.4.2
openai==2.29.0
packaging==24.0
platformdirs==4.2.0
propcache==0.4.1
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.17.2
pytest==7.4.4
python-dotenv==1.0.1
PyYAML==6.0.1
regex==2026.1.15
requests==2.31.0
rich==13.7.1
rlp==4.1.0
setuptools==68.1.2
sniffio==1.3.1
toolz==1.1.0
typing_extensions==4.15.0
urllib3==2.0.7
web3==7.14.1
websockets==15.0.1
wheel==0.42.0
yarl==1.22.0
